requirement.txt
easyocr
pillow
fastapi